backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
python-multipart==0.0.9
ultralytics>=8.3.0
opencv-python-headless>=4.10.0.84
Pillow>=10.4.0
numpy>=2.1.0
pyyaml>=6.0.2
aiofiles>=23.2.1
